Switched sha256 to size_t variables

+++ sha256.c
@@ -112,7 +112,7 @@ void sha256_init(SHA256_CTX* ctx) {
 	ctx->count = 0;
 }
 
-void sha256_update(SHA256_CTX* ctx, const uint8_t* message, uint32_t msize) {
+void sha256_update(SHA256_CTX* ctx, const uint8_t* message, size_t msize) {
 	while(msize > 0) {
 		const int bufoff = ctx->count % 64;
 		if(bufoff + msize <= 64) {
@@ -173,7 +173,7 @@ void sha256_final(SHA256_CTX* ctx, uint8_t sum[32]) {
 	memset(ctx, 0, sizeof(SHA256_CTX));
 }
 
-void sha256(const uint8_t* message, uint64_t osize, uint8_t* out) {
+void sha256(const uint8_t* message, size_t osize, uint8_t* out) {
 	SHA256_CTX ctx;
 	sha256_init(&ctx);
 	sha256_update(&ctx, message, osize);
@@ -182,7 +182,7 @@ void sha256(const uint8_t* message, uint64_t osize, uint8_t* out) {
 	/* no need to zero ctx as final does it */
 }
 
-void hmac_sha256_init(HMAC_SHA256_CTX* ctx, const uint8_t* _key, uint32_t keylen) {
+void hmac_sha256_init(HMAC_SHA256_CTX* ctx, const uint8_t* _key, size_t keylen) {
 	const uint8_t* key = _key;
 	uint8_t khash[32];
 	uint8_t pad[64];
@@ -213,7 +213,7 @@ void hmac_sha256_init(HMAC_SHA256_CTX* ctx, const uint8_t* _key, uint32_t keylen
 	memset(pad, 0, 64);
 }
 
-void hmac_sha256_update(HMAC_SHA256_CTX* ctx, uint8_t* message, uint32_t mlen) {
+void hmac_sha256_update(HMAC_SHA256_CTX* ctx, uint8_t* message, size_t mlen) {
 	/* feed the data into the inner context */
 	sha256_update(&(ctx->ictx), message, mlen);
 }
@@ -232,7 +232,7 @@ void hmac_sha256_final(HMAC_SHA256_CTX* ctx, uint8_t mac[32]) {
 	/* clean context unecessary as final was called on ictx and octx */
 }
 
-void hmac_sha256(uint8_t* key, uint32_t keylen, uint8_t* message, uint32_t len, uint8_t* out) {
+void hmac_sha256(uint8_t* key, size_t keylen, uint8_t* message, size_t len, uint8_t* out) {
 	HMAC_SHA256_CTX ctx;
 	hmac_sha256_init(&ctx, key, keylen);
 	hmac_sha256_update(&ctx, message, len);
@@ -242,7 +242,7 @@ void hmac_sha256(uint8_t* key, uint32_t keylen, uint8_t* message, uint32_t len, 
 // PBKDF2_HMAC_SHA256
 
 // dkLen and hlen are in bytes
-void pbkdf2_hmac_sha256(uint8_t* pass, uint32_t plen, uint8_t* salt, uint32_t saltLen, uint32_t c, uint32_t dkLen, uint8_t* out) {
+void pbkdf2_hmac_sha256(uint8_t* pass, size_t plen, uint8_t* salt, size_t saltLen, uint32_t c, size_t dkLen, uint8_t* out) {
 	/* in case dkLen is not a multiple of 32 */
 	const uint32_t sections = (dkLen + 31)/32;
 	/* 32 bit buffer to store integer counter */
